Move single-use code out of beatertest
... and into the package where it is used.
